% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eaf2maf.R
\name{eaf2maf}
\alias{eaf2maf}
\title{Transforms EAF to MAF
Turns effect allele frequency (between 0 and 1) into minor allele frequency (between 0 and 0.5)}
\usage{
eaf2maf(eaf)
}
\arguments{
\item{eaf}{effect allele frequecy, numeric between 0 and 1.}
}
\value{
P
}
\description{
Transforms EAF to MAF
Turns effect allele frequency (between 0 and 1) into minor allele frequency (between 0 and 0.5)
}
\examples{
eaf2maf(0.7)

}
