% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_nbr_tests.R
\name{eff_nbr_tests}
\alias{eff_nbr_tests}
\title{Effective number of tests}
\usage{
eff_nbr_tests(mat, cor.true = FALSE)
}
\arguments{
\item{mat}{matrix, either individual by SNP data, or correlation matrix of SNPs}

\item{cor.true}{indicates if option mat is a correlation matrix or not}
}
\value{
zhc.correction number of effective tests
}
\description{
Calculates the effective number of tests performed in a GWAS, given the LD structure among the SNPs. SNPs that are highly correlated are not counted twice.
}
\examples{

require(tidyverse)
dat.sim <- tibble::tibble(g = rnorm(100)) \%>\% mutate(g2 = g*2)
eff_nbr_tests(diag(10), cor.true = TRUE) ## 10, cause all independent
eff_nbr_tests(dat.sim, cor.true = FALSE)
}
