% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_inflation.R
\name{genomic_inflation}
\alias{genomic_inflation}
\title{Genomic inflation factor}
\usage{
genomic_inflation(Z = NULL, P = NULL)
}
\arguments{
\item{Z}{Numeric vector of arbitrary length. Z-statistics (sign does not matter)}

\item{P}{Numeric vector of arbitrary length. Alternatively, the P-values can be provided, which will be transformed into Z-statistics}
}
\value{
value
}
\description{
Genomic inflation factor
}
\examples{

data("giant")
genomic_inflation(Z = giant$BETA / giant$SE)
## should give the same as
genomic_inflation(P = giant$P)
}
