% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{giant}
\alias{giant}
\title{Height GWAS}
\format{The data set contains 10 variables:
\describe{
\item{\code{chr}}{}
\item{\code{pos}}{}
\item{\code{snp}}{}
\item{\code{tested_allele}}{}
\item{\code{other_allele}}{}
\item{\code{freq_tested_allele_hrc}}{}
\item{\code{beta}}{}
\item{\code{se}}{}
\item{\code{p}}{}
\item{\code{n}}{}
}

Full data: \url{}}
\usage{
giant
}
\description{
Associations between human body height and genetic markers.
Random selection of 10'000 SNPs.
}
\keyword{datasets}
