% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS.utils-package.R
\docType{package}
\name{GWAS.utils-package}
\alias{GWAS.utils}
\alias{GWAS.utils-package}
\title{GWAS.utils: Helper functions for GWAS data}
\description{
An R package with basic helper functions for manipulating GWAS data, including two GWAS datasets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sinarueeger/GWAS.utils}
  \item Report bugs at \url{https://github.com/sinarueeger/GWAS.utils/issues}
}

}
\author{
\strong{Maintainer}: Sina Rüeger \email{sina.rueeger@gmail.com}

Authors:
\itemize{
  \item Aurélien Macé
  \item Zoltán Kutalik
}

}
\keyword{internal}
