% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_normal.R
\name{inv_normal}
\alias{inv_normal}
\title{Inverse normal transformation}
\usage{
inv_normal(x)
}
\arguments{
\item{x}{numeric}
}
\value{
x.trans standard normal distributed
}
\description{
Transforms any variable into a standard normal distributed variable N(0,1).
}
\examples{
X <- runif(1000)

X.trans <- inv_normal(X)
par(mfrow = c(1,2))
hist(X)
hist(X.trans)
}
