% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QQplot.R
\name{QQplot}
\alias{QQplot}
\title{Pretty QQ plot}
\usage{
QQplot(p, neff = NULL, main = "", col = "black", add = FALSE, ...)
}
\arguments{
\item{p}{numeric vector, values between 0 and 1.}

\item{neff}{number of effective tests. see function \code{nbr_eff_tests}}

\item{main}{Main title}

\item{col}{color code}

\item{add}{logic, added to plot or not}

\item{...}{pass plot arguments on}
}
\description{
Creates a QQplot, given P-values
}
\examples{

QQplot(runif(100))
}
