% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_normal.R
\name{trans_inv_normal}
\alias{trans_inv_normal}
\title{Inverse normal transformation}
\usage{
trans_inv_normal(x, const = 3/8)
}
\arguments{
\item{x}{numeric vector that needs to be transformed}

\item{const}{numeric value, 3/8 and 0.5 recommended.}
}
\value{
Numeric vector of length(x) that is standard normally distributed
}
\description{
Transforms any variable into a standard normal distributed variable N(0,1).
}
\examples{
X <- runif(1000)

X.trans <- trans_inv_normal(X)
par(mfrow = c(1, 2))
hist(X)
hist(X.trans)
par(mfrow = c(1, 2))
qqnorm(X)
qqnorm(X.trans)
}
\references{
See also https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2921808/
}
