% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Z2P.R
\name{Z2P}
\alias{Z2P}
\title{Transform Z-statistic into P-value}
\usage{
Z2P(Z)
}
\arguments{
\item{Z}{Z-statistic (neg or pos), numeric vector of length 1 or longer.}
}
\value{
P P-value, same length as Z.
}
\description{
Transforms Z-statistic into P-value. Works for large Z-statistics too.
}
\examples{
Z2P(-1.959964)
Z2P(20)

}
