% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS.utils-package.R
\docType{package}
\name{GWAS.utils-package}
\alias{GWAS.utils}
\alias{GWAS.utils-package}
\title{GWAS.utils: GWAS specific functions}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/sinarueeger/GWAS.utils/issues}
}

}
\author{
\strong{Maintainer}: Sina R<U+00FC>eger \email{sina.rueeger@gmail.com}

Authors:
\itemize{
  \item Aur<U+00E9>lien Mac<U+00E9>
  \item Zolt<U+00E1>n Kutalik
}

}
\keyword{internal}
