% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QQplot.R
\name{QQplot}
\alias{QQplot}
\title{Pretty Q-Q-plot for P values}
\usage{
QQplot(p, neff = NULL, main = "QQplot", col = "black", add = FALSE,
  ...)
}
\arguments{
\item{p}{numeric vector, values between 0 and 1.}

\item{neff}{number of effective tests,
see function \code{\link{eff_nbr_tests}}}

\item{main}{Main title}

\item{col}{color code}

\item{add}{logic, added to plot or not}

\item{...}{pass plot arguments on}
}
\value{
plot (base)
}
\description{
Creates a QQplot, given P-values, which should be uniformly distributed
}
\examples{

QQplot(runif(100))
QQplot(runif(100), main = "This is the main title")
QQplot(runif(100), add = TRUE, col = "green")
}
\seealso{
\url{https://github.com/sinarueeger/ggGWAS} (in development)
and \code{\link[qqman]{qq}}
}
