% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z2p.R
\name{z2p}
\alias{z2p}
\title{Transforms Z-statistic into P-value}
\usage{
z2p(Z, method = c("pnorm"))
}
\arguments{
\item{Z}{Z-statistic, numeric vector.}

\item{method}{\code{"prnorm"} (default), \code{"Rmpfr::pnorm"} (from the Rmpfr package)}
}
\value{
P P-value, same length as Z.
}
\description{
Takes Z-statistic(s) as input and outputs P-value(s). This function works
for absolute Z-statistics up to x
}
\details{
The usual P value calc is \code{pnorm(abs(Z), lower = FALSE) * 2}. But for larger absolute
Z statistics (~ 40) we can use the log \code{exp(pnorm(abs(Z), log.p = TRUE, lower = FALSE)) * 2}.
For anything that is larger, we can use the R package \code{Rmpfr::pnorm} that helps us with small digits.
See also this post: https://stackoverflow.com/questions/46416027/how-to-compute-p-values-from-z-scores-in-r-when-the-z-score-is-large-pvalue-muc
}
\examples{
z2p(c(-1.959964, 1.96))
z2p(20)
z2p(20, method = "Rmpfr::pnorm")
z2p(200, method = "pnorm") # the default
z2p(200, method = "Rmpfr::pnorm")



}
